backend/requirements.txt
fastapi
uvicorn
sqlalchemy>=2.0
asyncpg
psycopg2-binary
python-dotenv
pydantic>=2